refactor(ARCHITECT-002): split AdminPrinters.jsx into sub-components
Extract PrinterModal, IPProbeSection, MaintenanceModal, and shared
constants from AdminPrinters.jsx (1668 → 964 lines, 42% reduction).
Uses static Tailwind class maps for MAINTENANCE_TYPE_CLASS to avoid
CSS purge issues in production builds.
