+++ frontend/src/pages/admin/AdminPrinters.jsx
@@ -1,25 +1,16 @@
+/**
+ * AdminPrinters - Printer fleet management orchestrator.
+ *
+ * Sub-components extracted per ARCHITECT-002:
+ *   PrinterModal, IPProbeSection, MaintenanceModal, constants
+ */
 import { useState, useEffect } from "react";
 import { API_URL } from "../../config/api";
 import { useToast } from "../../components/Toast";
-import Modal from "../../components/Modal";
-
-const statusColors = {
-  offline: "bg-gray-500/20 text-gray-400",
-  idle: "bg-green-500/20 text-green-400",
-  printing: "bg-blue-500/20 text-blue-400",
-  paused: "bg-yellow-500/20 text-yellow-400",
-  error: "bg-red-500/20 text-red-400",
-  maintenance: "bg-orange-500/20 text-orange-400",
-};
-
-const brandLabels = {
-  bambulab: "BambuLab",
-  klipper: "Klipper/Moonraker",
-  octoprint: "OctoPrint",
-  prusa: "Prusa",
-  creality: "Creality",
-  generic: "Generic/Manual",
-};
+import { statusColors, brandLabels, MAINTENANCE_TYPE_CLASS } from "../../components/printers/constants";
+import PrinterModal from "../../components/printers/PrinterModal";
+import IPProbeSection from "../../components/printers/IPProbeSection";
+import MaintenanceModal from "../../components/printers/MaintenanceModal";
 
 export default function AdminPrinters() {
   const toast = useToast();
@@ -897,10 +888,7 @@ export default function AdminPrinters() {
                         </td>
                         <td className="py-3 px-4">
                           <span className={`px-2 py-1 rounded-full text-xs capitalize ${
-                            log.maintenance_type === "repair" ? "bg-red-500/20 text-red-400" :
-                            log.maintenance_type === "routine" ? "bg-green-500/20 text-green-400" :
-                            log.maintenance_type === "calibration" ? "bg-blue-500/20 text-blue-400" :
-                            "bg-purple-500/20 text-purple-400"
+                            MAINTENANCE_TYPE_CLASS[log.maintenance_type] || "bg-purple-500/20 text-purple-400"
                           }`}>
                             {log.maintenance_type}
                           </span>
@@ -974,694 +962,3 @@ export default function AdminPrinters() {
   );
 }
 
-// ============================================================================
-// Printer Modal Component
-// ============================================================================
-
-function PrinterModal({ printer, onClose, onSave, brandInfo }) {
-  const toast = useToast();
-  const [loading, setLoading] = useState(false);
-  const [workCenters, setWorkCenters] = useState([]);
-  const [form, setForm] = useState({
-    code: printer?.code || "",
-    name: printer?.name || "",
-    model: printer?.model || "",
-    brand: printer?.brand || "generic",
-    serial_number: printer?.serial_number || "",
-    ip_address: printer?.ip_address || "",
-    access_code: printer?.connection_config?.access_code || "",
-    location: printer?.location || "",
-    work_center_id: printer?.work_center_id || "",
-    notes: printer?.notes || "",
-    active: printer?.active !== false,
-  });
-
-  const token = localStorage.getItem("adminToken");
-  const isEdit = !!printer;
-
-  // Fetch machine-type work centers on mount
-  useEffect(() => {
-    const fetchWorkCenters = async () => {
-      try {
-        const res = await fetch(`${API_URL}/api/v1/work-centers/?center_type=machine`, {
-          headers: { Authorization: `Bearer ${token}` },
-        });
-        if (res.ok) {
-          const data = await res.json();
-          setWorkCenters(data);
-        }
-      } catch {
-        // Non-critical - work center selection is optional
-      }
-    };
-    fetchWorkCenters();
-  }, [token]);
-
-  // Get models for selected brand
-  const selectedBrand = brandInfo.find((b) => b.code === form.brand);
-  const models = selectedBrand?.models || [];
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      const url = isEdit
-        ? `${API_URL}/api/v1/printers/${printer.id}`
-        : `${API_URL}/api/v1/printers`;
-
-      // Build payload with connection_config for brand-specific settings
-      const { access_code, ...rest } = form;
-      const payload = {
-        ...rest,
-        connection_config: access_code ? { access_code } : {},
-      };
-
-      const res = await fetch(url, {
-        method: isEdit ? "PUT" : "POST",
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(payload),
-      });
-
-      if (!res.ok) {
-        const err = await res.json();
-        throw new Error(err.detail || "Failed to save printer");
-      }
-
-      toast.success(isEdit ? "Printer updated" : "Printer added");
-      onSave();
-    } catch (err) {
-      toast.error(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const generateCode = async () => {
-    try {
-      const prefix = form.brand === "generic" ? "PRT" : form.brand.toUpperCase().slice(0, 3);
-      const res = await fetch(`${API_URL}/api/v1/printers/generate-code?prefix=${prefix}`, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-      if (res.ok) {
-        const data = await res.json();
-        setForm({ ...form, code: data.code });
-      }
-    } catch {
-      // Non-critical
-    }
-  };
-
-  return (
-    <Modal
-      isOpen={true}
-      onClose={onClose}
-      title={isEdit ? "Edit Printer" : "Add Printer"}
-      className="w-full max-w-lg max-h-[90vh] overflow-y-auto"
-      disableClose={loading}
-    >
-      <div className="p-6 border-b border-gray-700">
-        <h2 className="text-xl font-bold text-white">
-          {isEdit ? "Edit Printer" : "Add Printer"}
-        </h2>
-      </div>
-
-      <form onSubmit={handleSubmit} className="p-6 space-y-4">
-          {/* Brand */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Brand</label>
-            <select
-              value={form.brand}
-              onChange={(e) => setForm({ ...form, brand: e.target.value, model: "" })}
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              {Object.entries(brandLabels).map(([value, label]) => (
-                <option key={value} value={value}>{label}</option>
-              ))}
-            </select>
-          </div>
-
-          {/* Model */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Model *</label>
-            {models.length > 0 ? (
-              <select
-                value={form.model}
-                onChange={(e) => setForm({ ...form, model: e.target.value })}
-                required
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-blue-500"
-              >
-                <option value="">Select model...</option>
-                {models.map((m) => (
-                  <option key={m.value} value={m.value}>{m.label}</option>
-                ))}
-              </select>
-            ) : (
-              <input
-                type="text"
-                value={form.model}
-                onChange={(e) => setForm({ ...form, model: e.target.value })}
-                required
-                placeholder="e.g., Ender 3 V2"
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-              />
-            )}
-          </div>
-
-          {/* Code */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Code *</label>
-            <div className="flex gap-2">
-              <input
-                type="text"
-                value={form.code}
-                onChange={(e) => setForm({ ...form, code: e.target.value })}
-                required
-                placeholder="PRT-001"
-                className="flex-1 bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-              />
-              {!isEdit && (
-                <button
-                  type="button"
-                  onClick={generateCode}
-                  className="px-3 py-2 text-gray-400 hover:text-white border border-gray-700 rounded-lg"
-                >
-                  Auto
-                </button>
-              )}
-            </div>
-          </div>
-
-          {/* Name */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Name *</label>
-            <input
-              type="text"
-              value={form.name}
-              onChange={(e) => setForm({ ...form, name: e.target.value })}
-              required
-              placeholder="e.g., X1C Bay 1"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          {/* IP Address */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">IP Address</label>
-            <input
-              type="text"
-              value={form.ip_address}
-              onChange={(e) => setForm({ ...form, ip_address: e.target.value })}
-              placeholder="192.168.1.100"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          {/* Access Code - BambuLab only */}
-          {form.brand === "bambulab" && (
-            <div>
-              <label className="block text-sm text-gray-300 mb-1">LAN Access Code</label>
-              <input
-                type="text"
-                value={form.access_code}
-                onChange={(e) => setForm({ ...form, access_code: e.target.value })}
-                placeholder="8-digit code from printer"
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-              />
-              <p className="text-xs text-gray-500 mt-1">
-                Find this in your printer's network settings
-              </p>
-            </div>
-          )}
-
-          {/* Serial Number */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Serial Number</label>
-            <input
-              type="text"
-              value={form.serial_number}
-              onChange={(e) => setForm({ ...form, serial_number: e.target.value })}
-              placeholder="Optional"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          {/* Location */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Location</label>
-            <input
-              type="text"
-              value={form.location}
-              onChange={(e) => setForm({ ...form, location: e.target.value })}
-              placeholder="e.g., Farm A, Bay 1"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          {/* Work Center (Machine Pool) */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Machine Pool</label>
-            <select
-              value={form.work_center_id}
-              onChange={(e) => setForm({ ...form, work_center_id: e.target.value ? parseInt(e.target.value) : "" })}
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              <option value="">None</option>
-              {workCenters.map((wc) => (
-                <option key={wc.id} value={wc.id}>{wc.name}</option>
-              ))}
-            </select>
-          </div>
-
-          {/* Notes */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Notes</label>
-            <textarea
-              value={form.notes}
-              onChange={(e) => setForm({ ...form, notes: e.target.value })}
-              placeholder="Optional notes..."
-              rows={2}
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          {/* Active toggle */}
-          <div className="flex items-center gap-3">
-            <input
-              type="checkbox"
-              id="active"
-              checked={form.active}
-              onChange={(e) => setForm({ ...form, active: e.target.checked })}
-              className="w-4 h-4 rounded bg-gray-800 border-gray-700 text-blue-500 focus:ring-blue-500"
-            />
-            <label htmlFor="active" className="text-gray-300">Active</label>
-          </div>
-
-          {/* Actions */}
-          <div className="flex gap-3 pt-4 border-t border-gray-700">
-            <button
-              type="button"
-              onClick={onClose}
-              className="flex-1 px-4 py-2 text-gray-400 hover:text-white border border-gray-700 rounded-lg transition-colors"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              disabled={loading}
-              className="flex-1 bg-blue-600 hover:bg-blue-500 disabled:bg-blue-600/50 text-white px-4 py-2 rounded-lg transition-colors"
-            >
-              {loading ? "Saving..." : isEdit ? "Update" : "Add Printer"}
-            </button>
-          </div>
-      </form>
-    </Modal>
-  );
-}
-
-// ============================================================================
-// IP Probe Component - For discovering printers by IP address
-// ============================================================================
-
-function IPProbeSection({ token, onPrinterFound }) {
-  const toast = useToast();
-  const [ipAddress, setIpAddress] = useState("");
-  const [probing, setProbing] = useState(false);
-  const [probeResult, setProbeResult] = useState(null);
-
-  const handleProbe = async () => {
-    if (!ipAddress.trim()) {
-      toast.error("Please enter an IP address");
-      return;
-    }
-
-    // Basic IP validation
-    const ipRegex = /^(\d{1,3}\.){3}\d{1,3}$/;
-    if (!ipRegex.test(ipAddress.trim())) {
-      toast.error("Please enter a valid IP address");
-      return;
-    }
-
-    setProbing(true);
-    setProbeResult(null);
-
-    try {
-      const res = await fetch(
-        `${API_URL}/api/v1/printers/probe-ip?ip_address=${encodeURIComponent(ipAddress.trim())}`,
-        {
-          method: "POST",
-          headers: { Authorization: `Bearer ${token}` },
-        }
-      );
-
-      if (!res.ok) throw new Error("Probe failed");
-
-      const result = await res.json();
-      setProbeResult(result);
-
-      if (result.reachable) {
-        toast.success(`Found ${result.brand || "printer"} at ${ipAddress}`);
-        if (!result.already_registered) {
-          // Add to discovered list
-          onPrinterFound({
-            ip_address: result.ip_address,
-            brand: result.brand || "generic",
-            model: result.model,
-            name: result.suggested_name,
-            suggested_code: result.suggested_code,
-            already_registered: false,
-          });
-        }
-      } else {
-        toast.error(`No printer found at ${ipAddress}`);
-      }
-    } catch (err) {
-      toast.error("Failed to probe IP: " + err.message);
-    } finally {
-      setProbing(false);
-    }
-  };
-
-  return (
-    <div className="bg-gray-800 border border-gray-700 rounded-xl p-6">
-      <h2 className="text-lg font-medium text-white mb-2">Find Printer by IP</h2>
-      <p className="text-gray-400 text-sm mb-4">
-        Enter a printer's IP address to detect it. Works with BambuLab, Klipper, and OctoPrint.
-      </p>
-
-      <div className="flex gap-3">
-        <input
-          type="text"
-          value={ipAddress}
-          onChange={(e) => setIpAddress(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && handleProbe()}
-          placeholder="192.168.1.100"
-          className="flex-1 bg-gray-900 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-blue-500 font-mono"
-        />
-        <button
-          onClick={handleProbe}
-          disabled={probing}
-          className="bg-blue-600 hover:bg-blue-500 disabled:bg-blue-600/50 text-white px-6 py-2 rounded-lg transition-colors flex items-center gap-2"
-        >
-          {probing ? (
-            <>
-              <svg className="animate-spin h-5 w-5" viewBox="0 0 24 24">
-                <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" fill="none" />
-                <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z" />
-              </svg>
-              Probing...
-            </>
-          ) : (
-            <>
-              <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8.111 16.404a5.5 5.5 0 017.778 0M12 20h.01m-7.08-7.071c3.904-3.905 10.236-3.905 14.141 0M1.394 9.393c5.857-5.857 15.355-5.857 21.213 0" />
-              </svg>
-              Probe
-            </>
-          )}
-        </button>
-      </div>
-
-      {/* Probe Result */}
-      {probeResult && (
-        <div className={`mt-4 p-3 rounded-lg ${probeResult.reachable ? "bg-green-500/10 border border-green-500/30" : "bg-red-500/10 border border-red-500/30"}`}>
-          {probeResult.reachable ? (
-            <div className="space-y-2">
-              <div className="flex items-center gap-2">
-                <svg className="w-5 h-5 text-green-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
-                </svg>
-                <span className="text-green-400 font-medium">Printer Found!</span>
-              </div>
-              <div className="text-sm text-gray-300 space-y-1">
-                <div>Brand: <span className="text-white">{probeResult.brand || "Unknown"}</span></div>
-                <div>Open ports: <span className="text-white font-mono text-xs">{probeResult.ports_open?.map(p => `${p.port} (${p.service})`).join(", ") || "None"}</span></div>
-                {probeResult.already_registered && (
-                  <div className="text-yellow-400">
-                    Already registered as: {probeResult.existing_printer?.name} ({probeResult.existing_printer?.code})
-                  </div>
-                )}
-              </div>
-            </div>
-          ) : (
-            <div className="flex items-center gap-2 text-red-400">
-              <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-              </svg>
-              <span>No printer detected at this IP</span>
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Quick tip */}
-      <div className="mt-4 text-xs text-gray-500">
-        Tip: Check your router's DHCP client list to find printer IPs, or look in your printer's network settings.
-      </div>
-    </div>
-  );
-}
-
-// ============================================================================
-// Maintenance Modal Component
-// ============================================================================
-
-function MaintenanceModal({ printers, selectedPrinterId, onClose, onSave }) {
-  const toast = useToast();
-  const [loading, setLoading] = useState(false);
-  const [form, setForm] = useState({
-    printer_id: selectedPrinterId || "",
-    maintenance_type: "routine",
-    description: "",
-    performed_by: "",
-    performed_at: new Date().toISOString().slice(0, 16),
-    next_due_at: "",
-    cost: "",
-    downtime_minutes: "",
-    parts_used: "",
-    notes: "",
-  });
-
-  const token = localStorage.getItem("adminToken");
-
-  const maintenanceTypes = [
-    { value: "routine", label: "Routine Maintenance", description: "Regular scheduled maintenance" },
-    { value: "repair", label: "Repair", description: "Fixing a broken component" },
-    { value: "calibration", label: "Calibration", description: "Bed leveling, extrusion tuning" },
-    { value: "cleaning", label: "Cleaning", description: "Nozzle, bed, or general cleaning" },
-  ];
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!form.printer_id) {
-      toast.error("Please select a printer");
-      return;
-    }
-
-    setLoading(true);
-
-    try {
-      const payload = {
-        maintenance_type: form.maintenance_type,
-        description: form.description || null,
-        performed_by: form.performed_by || null,
-        performed_at: form.performed_at ? new Date(form.performed_at).toISOString() : new Date().toISOString(),
-        next_due_at: form.next_due_at ? new Date(form.next_due_at).toISOString() : null,
-        cost: form.cost ? parseFloat(form.cost) : null,
-        downtime_minutes: form.downtime_minutes ? parseInt(form.downtime_minutes) : null,
-        parts_used: form.parts_used || null,
-        notes: form.notes || null,
-      };
-
-      const res = await fetch(`${API_URL}/api/v1/maintenance/printers/${form.printer_id}/maintenance`, {
-        method: "POST",
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(payload),
-      });
-
-      if (!res.ok) {
-        const err = await res.json();
-        throw new Error(err.detail || "Failed to log maintenance");
-      }
-
-      toast.success("Maintenance logged successfully");
-      onSave();
-    } catch (err) {
-      toast.error(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <Modal
-      isOpen={true}
-      onClose={onClose}
-      title="Log Maintenance"
-      className="w-full max-w-lg max-h-[90vh] overflow-y-auto"
-      disableClose={loading}
-    >
-      <div className="p-6 border-b border-gray-700">
-        <h2 className="text-xl font-bold text-white">Log Maintenance</h2>
-        <p className="text-gray-400 text-sm mt-1">Track maintenance activities, costs, and downtime</p>
-      </div>
-
-      <form onSubmit={handleSubmit} className="p-6 space-y-4">
-          {/* Printer Selection */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Printer *</label>
-            <select
-              value={form.printer_id}
-              onChange={(e) => setForm({ ...form, printer_id: e.target.value })}
-              required
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-orange-500"
-            >
-              <option value="">Select printer...</option>
-              {printers.map((p) => (
-                <option key={p.id} value={p.id}>{p.name} ({p.code})</option>
-              ))}
-            </select>
-          </div>
-
-          {/* Maintenance Type */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Type *</label>
-            <select
-              value={form.maintenance_type}
-              onChange={(e) => setForm({ ...form, maintenance_type: e.target.value })}
-              required
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-orange-500"
-            >
-              {maintenanceTypes.map((t) => (
-                <option key={t.value} value={t.value}>{t.label}</option>
-              ))}
-            </select>
-          </div>
-
-          {/* Description */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Description</label>
-            <input
-              type="text"
-              value={form.description}
-              onChange={(e) => setForm({ ...form, description: e.target.value })}
-              placeholder="e.g., Replaced nozzle, cleaned bed"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-            />
-          </div>
-
-          {/* Performed By */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Performed By</label>
-            <input
-              type="text"
-              value={form.performed_by}
-              onChange={(e) => setForm({ ...form, performed_by: e.target.value })}
-              placeholder="Your name"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-            />
-          </div>
-
-          {/* Date/Time Row */}
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm text-gray-300 mb-1">Performed At *</label>
-              <input
-                type="datetime-local"
-                value={form.performed_at}
-                onChange={(e) => setForm({ ...form, performed_at: e.target.value })}
-                required
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-orange-500"
-              />
-            </div>
-            <div>
-              <label className="block text-sm text-gray-300 mb-1">Next Due</label>
-              <input
-                type="datetime-local"
-                value={form.next_due_at}
-                onChange={(e) => setForm({ ...form, next_due_at: e.target.value })}
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white focus:outline-none focus:ring-2 focus:ring-orange-500"
-              />
-            </div>
-          </div>
-
-          {/* Cost and Downtime Row */}
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm text-gray-300 mb-1">Cost ($)</label>
-              <input
-                type="number"
-                step="0.01"
-                min="0"
-                value={form.cost}
-                onChange={(e) => setForm({ ...form, cost: e.target.value })}
-                placeholder="0.00"
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-              />
-            </div>
-            <div>
-              <label className="block text-sm text-gray-300 mb-1">Downtime (minutes)</label>
-              <input
-                type="number"
-                min="0"
-                value={form.downtime_minutes}
-                onChange={(e) => setForm({ ...form, downtime_minutes: e.target.value })}
-                placeholder="0"
-                className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-              />
-            </div>
-          </div>
-
-          {/* Parts Used */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Parts Used</label>
-            <input
-              type="text"
-              value={form.parts_used}
-              onChange={(e) => setForm({ ...form, parts_used: e.target.value })}
-              placeholder="e.g., Hardened nozzle 0.4mm, PTFE tube"
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-            />
-            <p className="text-xs text-gray-500 mt-1">Comma-separated list of parts used</p>
-          </div>
-
-          {/* Notes */}
-          <div>
-            <label className="block text-sm text-gray-300 mb-1">Notes</label>
-            <textarea
-              value={form.notes}
-              onChange={(e) => setForm({ ...form, notes: e.target.value })}
-              placeholder="Additional notes..."
-              rows={2}
-              className="w-full bg-gray-800 border border-gray-700 rounded-lg px-4 py-2 text-white placeholder-gray-500 focus:outline-none focus:ring-2 focus:ring-orange-500"
-            />
-          </div>
-
-          {/* Actions */}
-          <div className="flex gap-3 pt-4 border-t border-gray-700">
-            <button
-              type="button"
-              onClick={onClose}
-              className="flex-1 px-4 py-2 text-gray-400 hover:text-white border border-gray-700 rounded-lg transition-colors"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              disabled={loading}
-              className="flex-1 bg-orange-600 hover:bg-orange-500 disabled:bg-orange-600/50 text-white px-4 py-2 rounded-lg transition-colors"
-            >
-              {loading ? "Saving..." : "Log Maintenance"}
-            </button>
-          </div>
-      </form>
-    </Modal>
-  );
-}
